Correcting one sideness conversion.
git-svn-id: https://www.enzo.desena.name/repos/mcl/trunk@19 6607b5ef-29ca-4f6a-8fd6-a5ad878f2eb4

+++ include/vectorop.h
@@ -145,7 +145,7 @@ std::vector<T> CircShift(const std::vector<T>& vector, Int num_positions) {
   UInt N = Length(vector);
   std::vector<T> output(N);
   for (UInt i=0; i<N; ++i) {
-    UInt index = Mod(((Int) i) - num_positions, (Int) N);
+    UInt index = (UInt) Mod(((Int) i) - num_positions, (Int) N);
     output[i] = vector[index];
   }
